+++ service/storage_service.hh
@@ -43,6 +43,7 @@
 
 class node_ops_cmd_request;
 class node_ops_cmd_response;
+struct node_ops_ctl;
 class node_ops_info;
 enum class node_ops_cmd : uint32_t;
 class repair_service;
@@ -90,7 +91,6 @@ class raft_group0;
 enum class disk_error { regular, commit };
 
 class node_ops_meta_data;
-struct node_ops_ctl;
 
 /**
  * This abstraction contains the token/identifier of this node
@@ -224,7 +224,7 @@ private:
         return _gossiper;
     };
 
-    friend struct node_ops_ctl;
+    friend struct ::node_ops_ctl;
 public:
 
     locator::effective_replication_map_factory& get_erm_factory() noexcept {
